Moved CredentialsSupplier to ServiceTransport
