+++ services-examples/src/main/java/io/scalecube/services/examples/auth/ServiceTransportAuthExample.java
@@ -3,9 +3,9 @@ package io.scalecube.services.examples.auth;
 import io.scalecube.services.Microservices;
 import io.scalecube.services.ServiceEndpoint;
 import io.scalecube.services.auth.Authenticator;
-import io.scalecube.services.auth.CredentialsSupplier;
 import io.scalecube.services.discovery.ScalecubeServiceDiscovery;
 import io.scalecube.services.exceptions.UnauthorizedException;
+import io.scalecube.services.transport.api.ServiceTransport.CredentialsSupplier;
 import io.scalecube.services.transport.rsocket.RSocketServiceTransport;
 import io.scalecube.transport.netty.websocket.WebsocketTransportFactory;
 import java.time.Duration;
