+++ services-transport-parent/services-transport-rsocket/src/main/java/io/scalecube/services/transport/rsocket/RSocketClientTransport.java
@@ -10,7 +10,6 @@ import io.rsocket.frame.decoder.PayloadDecoder;
 import io.rsocket.util.ByteBufPayload;
 import io.scalecube.services.Address;
 import io.scalecube.services.ServiceReference;
-import io.scalecube.services.auth.CredentialsSupplier;
 import io.scalecube.services.exceptions.MessageCodecException;
 import io.scalecube.services.exceptions.ServiceException;
 import io.scalecube.services.exceptions.UnauthorizedException;
@@ -18,6 +17,7 @@ import io.scalecube.services.transport.api.ClientChannel;
 import io.scalecube.services.transport.api.ClientTransport;
 import io.scalecube.services.transport.api.DataCodec;
 import io.scalecube.services.transport.api.HeadersCodec;
+import io.scalecube.services.transport.api.ServiceTransport.CredentialsSupplier;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
